Add asset pipeline using flask-assets library
Lando UI was missing a way to allow developers to write css and
javascript files and have those files all be compiled to one file which
would be served to the user on page load. Without this functionality
we would have to include every single css / javascript file by hand
in the html - which would be a strain on productivity.

Flask-Assets (https://flask-assets.readthedocs.io/en/latest/) is a
library which helps us solve this problem. It allows us to organize our
code in different files, write a manifest file defining how those files
should be compiled into one big file, and allows us to do some
pre/post processing on the files.

This commit orgainizes our css and javascript code underneath the
standard flask static folder. Inside is a src folder which contains
the js and css files as well as an 'assets.yml' file. The assets.yml
file is where we define how everything should be combined together.
Flask-Assets provides a Jinja template helper that will automatically
generate a link to the bundled asset when you want to include it in
the html.

This commit also initialize our asset pipeline with some defaults and
some examples. Notably Semantic-UI is included (along with jQuery which
it depends on). The home page has also been revamped to demonstrate how
to use the new asset pipeline.