Add support for building production static assets
In development, the development flask server serves static assets and
further it will recompile assets when assets have changed between
requests.

In production we don't want to do this. Ideally we want to compile the
static assets once and serve them always. This commit solves the two
problems needed to achive this:

1) This integrates flask-assets cli tools which exposes a command
   `flask assets build` which builds and outputs the assets. This
   command is run in the development image so that the production
   image can be free of build dependancies.

2) It updates the production image and circle.yml to copy in the
   static assets. These assets will remain unchanged until the next
   production image is built. This means that the assets could be
   served via proxy or volume mounted and served via nginx.

This also fixes the problem with the circle yml not pushing to
dockerhub correctly.