Simplify invoke tasks that run python unit tests.
The old invoke file was based on our invoke file from our conduit
hg monorepo - which had many different containers building and running
in unision.

In our case we're using invoke simply as an alias for docker-compose to
prevent us from having to type everything out.

The old invoke file created a new project space for the test containers
which was causing problems because changes to the Dockerfile or the
python requirements wouldn't cause rebuild of images in the testproject
space (unknown reason why). So what would happen is I would run
`invoke test` but I would get a build failure even though everything looked
fine on my end - the underlying cause was that the test container wasn't
refreshed and it was not obvious when this happened.

This commit changes the tasks.py file to use the same project space as
the normal development one. Now if the tests are failing it is because of
a problem that you should also have encountered when running the dev
containers.