Update tests to account for required SERVER_NAME setting
With the addition of the Auth0 changes, the Flask app now requires
that the SERVER_NAME config be set and be exactly the equal to the
host:port that the site is accessed by.

In the test environment that host:port is lando-ui:7777.
This commit updates the tests to set that and also updates the click
configuration to not set a default SESSION_COOKIE_DOMAIN since that
should always be set by whoever is creating the app.