Add /signin route that redirect to original location.
flask-pyoidc allows us to protect routes with the 'oidc_auth'
decorator which forces the user to login with Auth0 before they can see
that page. This 'oidc_auth' decorator "redirects" the user back to their
original page by detecting which view function it was called on.

Unfortunately this won't work for Lando because Lando should allow
users to view a page without logging in first, thus we cannot use
the 'oidc_auth' decorator.

This patch adds a helper function which logs the last valid page the
user was on at any point (normal http referrers won't work because that
includes the Auth0 routes). This also adds a signin route which is
protected, but, when accessed simply redirects the user to the last page
they were on, thus satisfying our login flow requirements.

An alternative option would be to add a '?redirect_url' param to the
login button. I opted to not do it this way because I thought it would
make the "signin" URL cluttered and also the last valid page session
variable can have other potential uses in the future.