Integrate Auth0 login into the UI navigation bar
This patch allows the user to click the big login button in the navbar
to login with Auth0. When logged in their name and picture will be
displayed. The user can click on their name to open up a settings modal
which will be used in the future to house all configurable settings.

This patch also adds support for a responsive navigation bar. When the
screen size is 1000px or smaller, a hamburger menu will appear to the
right instead of the usual buttons. When clicked the buttons will appear
in a new row below the original navigation bar.