revisions: remove 'D' prefixes to match new lando-api behavior
Summary:
lando-api changed in D255 so that all input and output to the api
concerning revisions will have the 'D' prefixed. This means that
the UI was now adding an extra 'D' in places were it used or displayed
the revision id.

This patch also updates the canned responses in the tests to match the
new expected api response.

Test Plan:
Run invoke test. You may also startup a local lando-api and
try visiting a revision from phabricator-dev.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Differential Revision: https://phabricator.services.mozilla.com/D303