revisions: properly redirect upon successful landing
Summary:
The _handle_submission method was attempting to redirect when a
landing is successful. Unfortunately, because the revisions_handler
method where it is called doesn't bubble up the return value, the
redirect is lost.

Prior to this commit, upon clicking the land button, even when successful
it would still appear that the revision has never been landed. This is
due to the fact that the redirect to the GET page (where the new landing
information would be loaded) never happends. This commit fixes the
behavior so that clicking the land button properly updates the page.

Note that this update is soley for the immediate success/failure to
request the landing, and it doesn't help with updating the asynchronous
state changes to the landing request if when the revision is either
successfully or unsuccessfully landed by transplant in the future.

Test Plan:
Run invoke test. If you want to test with a local lando-api
remove lines 34 and 38-44 in landoapi/api/landings.py to bypass auth.
Then change the return value in landoapi/transplant_client.py line 44
simulate a different unique (random works) id. Then you may click the
land button on a revision to try it out.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Differential Revision: https://phabricator.services.mozilla.com/D304