revisions: update landing button based on the revision landing status
Summary:
Updates the landing button to either be enabled/disabled/say landing is
queued depending on the revision's landing status.

Statuses can range from:
'aborted': This means something really bad happend, landoui should
"never" have to deal with this. In these cases the button is enabled.

'submitted': The landing has been queued in transplant and is in progress.
The button is disabled in this case when the page is loaded. When
 the user clicks on a enabled land button, this commit adds some
 javascript to immediately disabled the button and change the color
 and text. This gives the user immediate feedback on what they just
 did and was almost necessary because lando-api take a very long time
 to process the request before the page reloads. This also helps
 differentiate between 'Landing is Queueing" and "Landing Queued",
 the act of queueing can itself fail (resulting in the aborted status).

'landed': The landing was successful. The button is enabled explicitly
says "Re-land" in cases when users want to reland something. This
should be protected behind a warning in the future.

'failed': The landing failed. The button is enabled for the user to try again.

no status: The button is enabled for first landing.

Test Plan: See the attached video. Run invoke test. Test out with a local lando-api if wanted.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Differential Revision: https://phabricator.services.mozilla.com/D327