auth: Add 'lando' scope to auth0 request payload
Summary:
lando-api requires that clients request the user's explicit authorization
to make landings on their behalf. This is done by adding the 'lando'
scope to the auth0 request payload.

Test Plan:
Fairly straightforward, but, you may login in the UI,
go to /info, copy your access token, and then submit that to
lando-api with debug statements of the access token details.

Reviewers: smacleod

Reviewed By: smacleod

Differential Revision: https://phabricator.services.mozilla.com/D339