revisions: request revision info with specified diff
Summary:
When the user is visits /revisions/D123/99, they specifically wish to
view that revision in the context of the diff with id 99. Prior to this
it would always be in the context of the latest diff and the URL would
be re-written to match the latests diff id - meaning it was impossible
to view older diffs.

This change allows users to view older diffs and updates the canned
lando-api responses that support this change.

Test Plan:
invoke test and fire up the api and ui and visit D27 (from
the dev deployment of phabricator) which has 2 diffs 88 (latest) and 53.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Differential Revision: https://phabricator.services.mozilla.com/D400