revisions: refactor warnings and add warning if not viewing latest diff
Summary:
Moves the warning checks into their own helper method and adds a new
warning if the user is viewing an old diff of a revision when a newer
diff is available.

Test Plan: invoke test and try it in the UI with D27 from phabricator-dev. You'll need to be running D401 locally on your lando-api as this patch depends on changes made in the api.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Differential Revision: https://phabricator.services.mozilla.com/D402