Fix lando-api connection exception handling
Summary:
Allow exceptions in Lando API connection handling to bubble up to the
Sentry and Flask Debugger middleware.  This ensures that tracebacks end
up in the docker container logs, exceptions are captured by Sentry, and
that the Flask Debugger is opened on local dev systems when the app is
in debug mode.

Test Plan: see bug steps to reproduce

Reviewers: imadueme

Reviewed By: imadueme

Bug #: 1430642

Differential Revision: https://phabricator.services.mozilla.com/D390