revisions: display error page when diff and revision in url don't belong to each other.
Summary:
Catches the 400 error returned by lando api when you attempt to
view a diff of a revision when that diff doesn't belong to that revision.
This is different than the 404 that happens when the diff doesn't exist at
all.

Test Plan:
invoke test, try it out by visiting a revision as usual then
change the diff in the url to a diff that belongs to another revision.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Subscribers: smacleod

Differential Revision: https://phabricator.services.mozilla.com/D424