revisions: Add UI for landing blockers.
Summary:
Landing blockers prevent users from initiating a landing under some
situations (e.g. they do not have permissions to).

This commit adds the fuctionaility to display those blockers in the
UI. They are displayed similarly to warnings except in red and without
the ability to check them away.

When there are no blockers, the landing button and warnings are
visible as normal.

When one or more landing blockers are present, the landing button is
removed and no warnings are visible to the user, since it is impossible
to proceed with landing. When the blockers clear and the page is reload,
any warnings will be visible and will represent the latest warnings as
returned by lando-api.

When the user is not logged in, it is impossible to query lando-api
for a list of warnings or blockers (since authentication is required).
In this case, lando-ui displays a custom (non-bypassable) warning
informing the user that they must first log in to initiate a landing.
Once logged in they will see any real blockers or warnings as returned
by lando-api.

Reviewers: smacleod

Reviewed By: smacleod

Differential Revision: https://phabricator.services.mozilla.com/D519