revisions: add error message box in ui
Summary:
Sometimes a landing can fail immediately (instead of being posted back
as a landing status). We would like to communicate these immediate
failures on the UI.

This commit adds a box on the revisions page which will display any
`error` passed to the template.

This can be used for communicating such immediate failures in the future.

Test Plan:
Run invoke test. View screenshots. Feel free to run lando-ui
and change the code so that an error message is passed to the template,
this would allow you to test this on your own monitor/browser combination.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1437107

Differential Revision: https://phabricator.services.mozilla.com/D570