revisions: display error when failing to enqueue a landing
Summary:
When a user attempts to enqueue a landing and that fails, we
should immediately infom them of what happened and why.

This commit does just that by converting all failing conditions
into a new exception class: `LandingSubmissionError`. The UI
can then catch this one class and display it's message to the user.

Test Plan:
invoke test, view the video in the comments. Feel free
to try it yourself with a local lando-ui and lando-api.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1437107

Differential Revision: https://phabricator.services.mozilla.com/D573