landoapiclient: add post_landings_dryrun method
Summary:
Adds a new method: post_landings_dryrun to communication with landoapi's
POST /landings/dryrun endpoint.

This will return the dryrun result according to the format landoapi defines.

This commit also changes the post_landings method to be consistent in
raising an AssertionError if the LandoAPIClient wasn't given an
auth0_access_token. Callers should handle the case when the user is not
logged in and attempting to land or see dryrun information.

Test Plan: invoke test, try it out.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1437542

Differential Revision: https://phabricator.services.mozilla.com/D618