landoapiclient: check if error response has a status_code before use
Summary:
When the get_revision and get_landings methods were unable to connect
to landoapi, they would return a generic error message due to another
exception in the error processing code.

These method incorrectly assumed that all RequestsExceptions that have
a response also have a response.status_code, which is not true for
ConnectTimeout exceptions.

Fixed to check if the status code is present and added tests. Now if
we're unable to connect to lando api, the user sees a helpful message
saying so.

Test Plan: invoke test, try it out.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1439583

Differential Revision: https://phabricator.services.mozilla.com/D619