revisions: use active diff id if diff id in url is None
Summary:
When a user visits /revisions/D1 lando-ui queries lando-api to get the
revision information expecting to get back diff info for the most recent
diff. On the other hand if the user visits /revisions/D1/99, they will
see diff information for diff 99 specifically.

Part of loading the revision page includes querying the dryrun endpoint
which requires a diff id. This doesn't happen if you're not logged in.

These two points made this bug hard to hit when doing local development:
the UI would crash in the case that the user is logged in and visits a
url without requesting a specific diff.

This commit fixes this problem to default to the diff provided in the
requested revision info if there is none in the url. (If there is one
in the url, the revision info diff and the url diff would be equal
by nature of how the get_revision query is made).

Test Plan: invoke test, try it out!

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1437542

Differential Revision: https://phabricator.services.mozilla.com/D622