auth: refresh expired tokens on revisions page (Bug 1448888).
Summary:
We now properly refresh the auth tokens when visiting the revisions page
with an expired token. The solution is quite hacky, calling out to
either a decorated and undecorated request handler depending on if the
user has an auth session. This is done because the needed information
and methods to call to handle this auth ourselves are buried in
Flask-pyoidc, as "private" methods and variables.

Test Plan:
`invoke test` passes.

Configured auth0 token expiry times to be extremely low and tested
session expiry and refresh by manually running lando UI / API.
Sessions were automatically refreshed when visiting a revision page
and the browser experience appeared seamless.

Reviewers: mars

Reviewed By: mars

Bug #: 1448888

Differential Revision: https://phabricator.services.mozilla.com/D1033