Add an ability to set the cookie to insecure mode also in production.
Summary:
Conduit Suite is using production Lando UI by default. The suite is running on
HTTP and cookies need to stay insecure.

If `SESSION_COOKIE_SECURE` environment variable was set to `0` uwsgi was
reading it as a string - `"0"`. This is now changed to an integer.

Talisman is taking control of the `SESSION_COOKIE_SECURE` config and
sets it to `True` if the app is not in debug mode. This patch is making the
uwsgi to read the `DEBUG` environment variable and passes it to the app.
`create_app` function is then setting the `app.debug` field.

Also, we're now logging most of the session configs.

Test Plan:
Set suite to use Dockerfile-prod and check if it is possible to log in to
Lando UI
`docker-compose -f docker-compose.yml -f docker-compose.lando-ui.yml -f docker-compose.override.yml up --build lando-ui.test lando-ui tinyproxy`

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1467446

Differential Revision: https://phabricator.services.mozilla.com/D1675