revisions: hide warnings, blockers, and actions when landing is enqueued (Bug 1457534)
Summary:
When a landing is enqueue, the UI knows that the user shouldn't be able
to trigger another landing. Instead of having the API show this as a blocker
the UI should hide the warnings, blockers, and land button itself.
Not doing so causes confusion when triggering a landing as there are different
messages coming from the timeline and the bottom of the page.

Test Plan: Visual inspection on the UI with a revision that was enqueued.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1457534

Differential Revision: https://phabricator.services.mozilla.com/D2067