revisions: improve user experience of landing (Bug 1473795)
Summary:
Prior to this commit, all the warnings and blockers would appear
on page load; e.g. when users just loaded the page and didn't even
intend on landing it, when they are viewing a revision that already
landed in the past, and so on.

Rather than give users this info when they may not need it, we now
display the warnings and blockers only after the user clicks the land
button. This gives the user context - that these messages are related
to their attempt to land the revision.

They can click the land button again to follow through with landing
(assuming no blockers, warnings are checked, etc.)

Test Plan: Manually test with some revisions. Video in the bug.

Reviewers: smacleod, mars

Reviewed By: mars

Subscribers: mars

Bug #: 1473795

Differential Revision: https://phabricator.services.mozilla.com/D2070