revisions: improve commit message look and toggling (Bug 1473797)
Summary:
Prior to this commit the commit message section had poor UX:
- The first line was repeated in the summary and the full expand.
- The summary was black on a white background, while the expand
  was black on a grey background.
- Clicking the toggle button didn't change it's text.
- The toggle button moved when you click it.

This commit addresses these problems by using only the full commit
message. It is presented in preformatted text, but, all with black
on a white background (which matches the other sections). The full
body is hidden using css line height tricks, so we don't need to
explicitly split the first line from the body. The body is only hidden when
the message is more than 8 lines long. If less than 8 lines, the entire message
is displayed and the toggle button is hidden.

The toggle button is now a simple link which expands and changes it's
text to match the action it will perform (i.e. 'Show'/'Hide').

Test Plan: Visual inspection on a revision page

Reviewers: smacleod, mars

Reviewed By: mars

Subscribers: mars

Bug #: 1473797

Differential Revision: https://phabricator.services.mozilla.com/D2097