navbar: remove prerelease message at the top (Bug 1475590)
Summary:
Lando is no longer in prerelease so we should remove the notice at the top.
Also the quick implementation of the message was causing some UI problems
on mobile devices, removing this fixes those issues.

Test Plan:
Visual inspection in the UI with responsive design mode on
in your browser.

Reviewers: smacleod

Reviewed By: smacleod

Subscribers: smacleod

Bug #: 1475590

Differential Revision: https://phabricator.services.mozilla.com/D2127