revisions: linkify bugs, revision urls, and diff ids (Bug 1473787)
Summary:
The UI references relevant items like bugs, revisions, and diffs but,
because they are in plaintext only, it is a pain for users to look up
that item. This commit aims to link these items to the correct resource.

Two new application configuration entires are added: BUGZILLA_URL and
PHABRICATOR_URL, these don't come from the environment, but, instead are
dependant on the LANDO_API_URL that is set.

From here, we're able to link to the correct instance of the two services
using regex string substitution in 3 new template filters.

Test Plan: Try it out. See screenshot. Invoke test.

Reviewers: mars

Reviewed By: mars

Subscribers: mars

Bug #: 1473787

Differential Revision: https://phabricator.services.mozilla.com/D2229