revisions: autolink to the commit for successful landings (Bug 1473790)
Summary:
This adds a new template helper which is used to linkify the commit ids
of successful landings.

It does not try to guess what a commit id looks like but instead matches
exactly the id given in the landing status.

Test Plan: invoke test and try it yourself in the UI!

Reviewers: mars, smacleod

Reviewed By: mars

Subscribers: mars

Bug #: 1473790

Differential Revision: https://phabricator.services.mozilla.com/D2297