oidc: add error handler view to fix session refresh and other errors (Bug 1474392)
Summary:
When something goes wrong with authentication, Auth0 redirects to our
provided redirect uri (simply /redirect_uri when using flask_pyoidc) with
two query parameters: error and error_description. Part of the session
refresh flow is to catch the error when the user's Single Sign On session
is expired, which we haven't been doing up to now.

Now, we hook into this using the @oidc.error_view decorator so we can handle
recoverable errors and log and show a better looking error message for
unrecoverable errors.

Test Plan:
1. _Without_ this revision, have lando-ui and lando-api up and running normally.
2. In your Auth0 account settings, set the Token Expiration Seconds for both the Auth0 Management API and lando-api api to something small like 10 seconds.
3. In landoui/auth.py set the session_refresh_interval_seconds to the same small time. This will effectively reproduce the problem.
4. Log out and log in to lando-ui so you have a fresh session.
5. Wait 10 seconds or the amount of time you set previously.
6. Visit a revision - you will get the old ugly error.
7. Apply this revision, keeping the change in auth.py.
8. Now view the revision, you will see that you are logged out and are able to login again normally.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1474392

Differential Revision: https://phabricator.services.mozilla.com/D2324