dockerfile: remove npm and eslint from dev dockerfile
Summary:
We were using npm and eslint when lando-ui was a react project.
After conversion to the flask app we stopped running eslint and so
it's not really needed in the dockerfile as of this commit.
Recently, it has been causing build failures and so we remove it
here for now until we properly add it back at some time in the future.

Reviewers: smacleod

Reviewed By: smacleod

Differential Revision: https://phabricator.services.mozilla.com/D2433