stacks: add initial UI for landing partial stacks (Bug 1473128).
Summary:
This change adds a completely new UI for viewing and landing Phabricator
revision stacks. Visiting '/D<revision-id>' will load the new UI, which
displays information about the revision stack D<revision-id> is part of.

This initial version displays a timeline like the old UI, but has
any request to transplant that includes a revision in the stack. Below
this timeline a table for revisions in the stack is displayed. At this
time the table will either include the landable ancestors in the stack
leading up to D<revision-id>, or just D<revision-id> itself if it is not
landable. A future change will extend this to display all revisions in
the stack, and allow choosing a specific landable series to land (This
is being iterated on as drawing DAGs is non trivial).

The stack table is designed to have a fixed height per row to make the
DAG drawing in the future much easier and emulate the stack view in
Phabricator itself. It will horizontally scroll if it requires extra
space (such as when there are MANY reviewers for a single revision).

Commit messages have been left out of this table and are only displayed
when attempting to preview the landing. This landing preview takes the
form of a modal when clicking the Preview Landing button. This modal is
where more information around warnings or things preventing a specific
user from landing the changes are displayed.

This UI was purposefully designed as a completely separate page and flow
to allow series support to be iterated on while Lando still operates as
usual. This first commit introduces only the display portion of landing
a series, a followup will include the ability to actually trigger a
landing.

Along with this change is a new client for Lando API,
landoui.landoapi.LandoAPI. This was added to consolidate a lot of the
error handling code needed when talking to Lando API, but was created
from scratch to avoid extra changes to the current UI. Going forward all
new code should make use of this client and when the original UI has
been removed so should it.

Depends on D3368.

Test Plan:
`invoke test` passes, new tests added. A lot of manually messing around
with various revisions from phabricator-dev, along with testing with
an exported lando-api database.

Reviewers: glob

Reviewed By: glob

Bug #: 1473128

Differential Revision: https://phabricator.services.mozilla.com/D3372