landing: Add an ability to see and land secure revisions.
Summary:
Phabricator API endpoints require a token to authenticate the user.

Currently, we use a default unprivileged token for all the requests. It gives us access to public revisions.
To be able to view and land secure revisions, Lando needs to use a token generated by the Phabricator user.

With this patch, Lando UI has a way for the authenticated user to provide the token (copy & paste using the browser).
It is then storing it in a secure, HttpOnly cookie (`phabricator-api-token`) and passes to the Lando API as the
`X-Phabricator-API-Key` header in every request. Lando UI is removing the token on user's demand or if the user
is logged out. Reusing the same token next time is possible.

The token is sent to the server with a CSRF protected POST request.

Test Plan: Attempt to land a secure revision.

Reviewers: imadueme, smacleod

Reviewed By: imadueme, smacleod

Subscribers: mcote, imadueme, zalun, keeler, Alex_Gaynor, mconley, sheehan, jkt, smacleod

Bug #: 1443704

Differential Revision: https://phabricator.services.mozilla.com/D2561