circleci: use multi-stage docker container to build static assets.
Summary:
In order to keep the prod image as small as possible, we can install ruby
and sass in an initial docker stage, and then copy it into the actual final
build stage of the dockerfile.

This required re-organizing the dockerfile -- as long as the steps in the
initial stage (labeled "assets") are the same as the subsequent stage, docker
will make use of the cache and there is no additional time required for the
build.

The downside of this is that we have to repeat ourselves a bit in the
dockerfile.

The version.json file is required to run the flask assets pipeline and so
I have moved its generation into the dockerfile. The CIRCLE_* values are passed
in using docker build-args (ARG).

Reviewers: smacleod

Reviewed By: smacleod

Subscribers: imadueme

Bug #: 1497520

Differential Revision: https://phabricator.services.mozilla.com/D8003