template_helpers: linkify the FAQ keyword (Bug 1496880)
Summary:
Some warnings and blocker message from the api have the FAQ keyword
in them. It would be good to turn that into a link the user can easily
click on, which we do in this commit.

Test Plan:
invoke test. you can simulate an invalid user by setting the
'LOCALDEV_MOCK_TRANSPLANT_SUBMIT' environment variable to 'inject_invalid'.

Reviewers: mars

Reviewed By: mars

Bug #: 1496880

Differential Revision: https://phabricator.services.mozilla.com/D10553