errorhandlers: pass through 301 redirect exceptions
Summary:
werkzeug 0.14 uses exceptions to signal a 301 redirect. before this commit
werkzeug worked differently and our catch all Exception handler was fine.
After upgrading it's clear that some exceptions need to be handled by the
flask app itself, such as this one.

Test Plan:
try going to a route without a / with and without this change.
Also run `invoke test` with and without this change.

Reviewers: dylan

Reviewed By: dylan

Subscribers: mcote

Bug #: 1504857

Differential Revision: https://phabricator.services.mozilla.com/D10978