errorhandlers: replace generic Exception handler with 500 code handler (bug 1507329)
Summary:
Flask knows how to handle some Exceptions, but, we were catching those
without giving the flask app a chance to handle them appropriately.
In the past there was some confusion, since unit tests would fail when
using a 500 code handler, but, pass when explictly catching all Exceptions.
The catch all Exception handler was left, but, then we were having to
manually determine which exceptions to catch vs which to pass back to
flask.

Here we use the proper 500 code handler which will catch all exceptions
that flask doesn't know how to handle, and we update the unit tests to
disable the TESTING and DEBUG flags when testing this behavior. When those
flags are set, flask doesn't handle any exceptions - useful for a dev or
testing enviroment, but, not when testing that exact scenario!

See http://flask.pocoo.org/docs/1.0/patterns/errorpages/#error-handlers

Test Plan: invoke test

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1507329

Differential Revision: https://phabricator.services.mozilla.com/D12061