landing preview: extend display of commit messages (Bug 1500223).
Summary:
Previously, long commit messages were cut off and the overflow was
hidden. This fix removes the max-height limitation to display the
commit message in full. The overflow property was removed as well,
as it is not necessary with this fix.

Test Plan:
manually ran lando-ui with a mock revision containing
long commit message and saw it was fully displayed.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1500223

Differential Revision: https://phabricator.services.mozilla.com/D12705