stacks: add algorithm for drawing the stack graph (Bug 1490337).
Summary:
This change introduces an algorithm for sorting the stack graph
topologically, implemented by `sort_stack_topological()`. It's important
to sort the stack topologically in order to draw it vertically beside
the revisions table.

`sort_stack_topological()` takes a `key` kwarg which can be used to
influence which topological sort is chosen. The intention here is to
allow us to influence the sort with `key` to result in a deterministic
order for a particular stack. I plan to break the tie between node
ordering using the revision ID in the future.

Additionally a `draw_stack_graph()` function implements an algorithm for
converting the topologically sorted graph into metadata which makes
drawing the graph trivial. A list of rows will be returned specifying
where the commit should appear in the drawing as well as the lines
needed to be drawn for the row. Each column specified in `above`,
`below`, or `other` results in a single line drawn. The specifics of
these lines is described in code.

Test Plan: `invoke test` passes with new tests added for the graph algorithms.

Reviewers: dylan

Reviewed By: dylan

Bug #: 1490337

Differential Revision: https://phabricator.services.mozilla.com/D14267