stacks: draw stack graph and allow selecting tip (Bug 1490337).
Summary:
This converts the current stack page to display the full stack graph
rather than just the revisions to be landed (or a single revision when
it can't land). The new stack graph drawn in the table indicates the
relationships between the rows and mimics the stack display in
Phabricator itself.

What can/will land is now indicated by the "Land" column. If a revision
is part of a landable path it will have a radio button in this column
allowing the user to select it as the tip to land. The currently
selected series that will land based on the tip has the "Land" column
highlighted green.

Due to time constraints dynamically changing the series to land based on
these radio button selections was scrapped. Instead clicking a radio
button will take the user to the stack page for that specific revision.
This results in a bit of a slow, sub-optimal UX, but gives the same
overall functionality. Because of this choice all the code for the
LandingPreview could stay the same, treating the selected series as
static.

In the future it would be good to transition to a more dynamic page
which would make the dryrun at selection time. This would require adding
some new dynamic endpoints the JS could hit to grab the dryrun as well
as restructuring the entire stack page's JS.

Depends on D14267.

Test Plan:
`invoke test` passes. Manually tested the new UI using various stacks
both simple and complex.

Reviewers: dylan

Reviewed By: dylan

Bug #: 1490337

Differential Revision: https://phabricator.services.mozilla.com/D14268