css: add delay to blocker tooltip to make it easy to click inside (bug 1515042)
Summary:
Before this commit, when a user hovers out of the blocker tooltip,
the tooltip would instantly close - making it difficult to click helpful
links within.

This commit adds a little animation and a delay that lets the tooltip
stay on the screen for a short time when the user's mouse moves outside it
before disappearing.

Test Plan: see gif below, try it out yourself

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1515042

Differential Revision: https://phabricator.services.mozilla.com/D15506