landing preview: separate toggleAll button into expand and collapse (Bug 1509515).
Summary:
Further to the last fix for Bug 1500223, which removed the max-height
limit and overflow properties; this fix improves the UI with the
addition of buttons to toggle the expansion and contraction of
individual, or all, long commit messages.

Below each commit message, the number of lines that are hidden is
displayed. A separate expandAllButton and collapseAllButton for all
commit messages provides a neat solution in which the user can display
or hide all long messages. Alternatively, each long message has its
own expand/collapse button with which the user can toggle individually
as they please.

Depends on D14996.

Test Plan: Manually tested in browser.

Reviewers: smacleod

Reviewed By: smacleod

Bug #: 1509515

Differential Revision: https://phabricator.services.mozilla.com/D15000