queue: add initial queue of landings on home page (bug 1587138)
Adds the initial queue of landings on the index page.

This queue will only show up when the ENABLE_EMBEDDED_TRANSPLANT_UI
flag is enabled and the user is logged in.
- In the future we can remove both of these requirements.

The landing queue currently fakes data from the api. Once the API
changes are ready we will update the UI to actually query real queue
data. The expected data format can be seen in `queue.html`.

Clicking on a revision will take the user to the revision page where
they will be able to cancel, prioritize, and see extra details about the
landing. These features will be implemented in a separate PR.