docker: fix sass builds
New version of webassets and flask-assets (2.0) is not compatible with
sass compiling with using the SCSS filter. New versions of sass require
the `--scss` flag to compile SCSS using STDIN, rather than the old
behaviour of defaulting to SCSS unless `--indented` is passed.

In order to upgrade webassets, flask-assets, and sass, we should either
move to the indented format, or look further into fixing the above
mentioned issue.