feat: implement Treestatus UI (Bug 1817473)
Implement a Treestatus UI in Lando. All endpoints in the new
interface are namespaced under `/treestatus`. The main components
are as follows.

The main Treestatus page displays all trees and their current statuses.
When authenticated, each tree can be selected for updating via a checkbox,
or a "select all trees" button is available for CI-wide tree closures.
The tree selections are forwarded to the update form when "Update trees"
is selected.

The "update trees" view implements a form for updating the status of
trees. Trees can be set to open, closed and "approval required". When
the status is being set to a non-open state, a reason and reason
category must be specified. The form presents a "remember this change"
option which can be used to easily revert the change at a later time.
For example if all trees are closed, the trees can be mass re-opened
easily.

The individual log view of a tree presents the most recent status
updates for each individual tree. When authenticated, users can edit
the tree status reason or category in case of a typo or error during
submission by clicking the "edit" button, which toggles the display to a
form for updating.

The recent changes view is a header that appears for authenticated users
on all pages. It presents all status updates that have been "remembered"
during update. Users can edit the reason and reason category using an
"edit" button that toggles the display to a form. The "restore" button
can be used to revert the status change. The "discard" button can be
used to throw away the status change if it is no longer necessary to
remember the change.

The Treestatus UI uses message flashing and the existing Lando-UI
notifications component to present errors and action verification to the
user.