stack: move the uplift request form into a separate modal (Bug 1801959)
Create a new modal div that contains the uplift request form. Add a new
button for requesting uplifts next to the "Preview Landing" button.
Remove the code that allows pressing the "Preview Landing" button
even when the stack is not landable since we no longer have the
uplift button behind that modal.